fix abnormal icon display

+++ media/base/ohos/ohos_media_player_bridge.cc
@@ -240,7 +240,7 @@ base::TimeDelta OHOSMediaPlayerBridge::GetMediaTime() {
   int32_t ret = player_->GetCurrentTime(time);
   if (ret == 0 && time == -1) {
     //if is livestream, return system time
-    time = base::Time::Now().ToInternalValue() / base::Time::kMicrosecondsPerMillisecond;
+    time = static_cast<int32_t>(base::Time::Now().ToInternalValue() / base::Time::kMicrosecondsPerMillisecond);
   }
   return base::Milliseconds(time);
 }
